add getHiddenFields util
