+++ packages/dataviews/src/utils/get-hidden-fields.ts
@@ -1,0 +1,23 @@
+/**
+ * Internal dependencies
+ */
+import type { NormalizedField, View } from '../types';
+
+export default function getHiddenFields< Item >(
+	view: View,
+	fields: NormalizedField< Item >[]
+): NormalizedField< Item >[] {
+	const visibleFieldIds = view.fields || [];
+	const togglableFields = [
+		view?.titleField,
+		view?.mediaField,
+		view?.descriptionField,
+	].filter( Boolean );
+	return fields.filter(
+		( f ) =>
+			! visibleFieldIds.includes( f.id ) &&
+			! togglableFields.includes( f.id ) &&
+			f.type !== 'media' &&
+			f.enableHiding !== false
+	);
+}
